+++ behaviors/index.js
@@ -30,3 +30,4 @@ export { initLegend } from './legend.js';
 export { initConnectors } from './connectors.js';
 export { initSpotlight } from './spotlight.js';
 export { initCrosshair } from './crosshair.js';
+export { initCommand } from './command.js';